halo-face-search/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Face recognition and processing
deepface==0.0.79
opencv-python==4.8.1.78
tensorflow==2.13.0

# Vector database
pymilvus==2.3.3

# Data processing
numpy==1.24.3
pandas==2.1.3
Pillow==10.1.0

# Caching and performance
redis==5.0.1
aioredis==2.0.1
fastapi-cache2==0.2.1

# Async support
aiofiles==23.2.1
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# StyleGAN dependencies (for face generation)
torch==2.1.1
torchvision==0.16.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2 
requests==2.31.0 